requirements.txt
astroid==2.4.2
azure-core==1.8.2
azure-data-tables==12.0.0b2
azure-functions==1.4.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
isodate==0.6.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
msrest==0.6.19
oauthlib==3.1.0
pylint==2.6.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
